Refactored data module
